requirements.txt
google-cloud-aiplatform==1.38.0
google-cloud-vision==3.4.5
google-cloud-speech==2.22.0
google-cloud-bigquery==3.12.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
vertexai==1.38.0
firebase-admin==6.2.0
langchain-google-vertexai==1.0.1
langchain==0.0.350
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
